+++ code/app/src/androidTest/java/com/cmput301f21t26/habittracker/ProfileFragmentTest.java
@@ -10,6 +10,8 @@ import android.util.Log;
 
 import androidx.test.rule.ActivityTestRule;
 
+import com.cmput301f21t26.habittracker.ui.LoginSignupActivity;
+
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
