+++ code/app/src/main/java/com/cmput301f21t26/habittracker/interfaces/UserCallback.java
@@ -1,0 +1,7 @@
+package com.cmput301f21t26.habittracker.interfaces;
+
+import com.cmput301f21t26.habittracker.objects.User;
+
+public interface UserCallback {
+    void onCallback(User user);
+}
