+++ code/app/src/main/java/com/cmput301f21t26/habittracker/objects/Habit.java
@@ -1,10 +1,7 @@
 package com.cmput301f21t26.habittracker.objects;
 
 import java.io.Serializable;
-import java.lang.reflect.Array;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.UUID;
