--- code/app/src/main/java/com/cmput301f21t26/habittracker/objects/UserCallback.java
@@ -1,5 +1,0 @@
-package com.cmput301f21t26.habittracker.objects;
-
-public interface UserCallback {
-    void onCallback(User user);
-}
