+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/auth/MainLoginSignupFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.fragments;
+package com.cmput301f21t26.habittracker.ui.auth;
 
 import android.os.Bundle;
