+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/auth/SignupFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.fragments;
+package com.cmput301f21t26.habittracker.ui.auth;
 
 import android.net.Uri;
 import android.os.Bundle;
