+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/habit/AddHabitFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.fragments;
+package com.cmput301f21t26.habittracker.ui.habit;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -19,11 +19,10 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
-import com.cmput301f21t26.habittracker.MainActivity;
+import com.cmput301f21t26.habittracker.ui.MainActivity;
 import com.cmput301f21t26.habittracker.R;
 import com.cmput301f21t26.habittracker.databinding.FragmentAddHabitBinding;
 import com.cmput301f21t26.habittracker.objects.Habit;
-import com.cmput301f21t26.habittracker.objects.User;
 import com.cmput301f21t26.habittracker.objects.UserController;
 import com.google.android.material.chip.Chip;
 import com.google.android.material.chip.ChipGroup;
