+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/habit/EditHabitFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.fragments;
+package com.cmput301f21t26.habittracker.ui.habit;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
@@ -21,7 +21,7 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
-import com.cmput301f21t26.habittracker.MainActivity;
+import com.cmput301f21t26.habittracker.ui.MainActivity;
 import com.cmput301f21t26.habittracker.R;
 import com.cmput301f21t26.habittracker.databinding.FragmentEditHabitBinding;
 import com.cmput301f21t26.habittracker.objects.Habit;
