+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/habit/TodayHabitFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.ui.home;
+package com.cmput301f21t26.habittracker.ui.habit;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,7 +13,7 @@ import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.cmput301f21t26.habittracker.HabitAdapter;
+import com.cmput301f21t26.habittracker.ui.habit.HabitAdapter;
 import com.cmput301f21t26.habittracker.R;
 import com.cmput301f21t26.habittracker.databinding.FragmentTodayHabitBinding;
 import com.cmput301f21t26.habittracker.objects.Habit;
