+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/habitevent/EditHabitEventFragment.java
@@ -1,8 +1,6 @@
-package com.cmput301f21t26.habittracker.fragments;
-
+package com.cmput301f21t26.habittracker.ui.habitevent;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -10,7 +8,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -20,32 +17,37 @@ import androidx.navigation.NavController;
 import androidx.navigation.NavDirections;
 import androidx.navigation.Navigation;
 
-import com.cmput301f21t26.habittracker.MainActivity;
+import com.cmput301f21t26.habittracker.ui.MainActivity;
 import com.cmput301f21t26.habittracker.MobileNavigationDirections;
 import com.cmput301f21t26.habittracker.R;
+import com.cmput301f21t26.habittracker.databinding.FragmentEditHabitEventBinding;
 import com.cmput301f21t26.habittracker.objects.Habit;
 import com.cmput301f21t26.habittracker.objects.HabitEvent;
+import com.cmput301f21t26.habittracker.objects.UserController;
+import com.google.android.material.textfield.TextInputEditText;
 
 import java.text.SimpleDateFormat;
 import java.util.Locale;
 
+public class EditHabitEventFragment extends Fragment {
 
-public class ViewHabitEventFragment extends Fragment {
-    private String TAG = "ViewHabitEventFragment";
+    private final String TAG = "EditHabitEventFragment";
 
+    private FragmentEditHabitEventBinding binding;
     private NavController navController;
-    private Button confirmHabitEventButton;
-    private Button editHabitEventButton;
-    private TextView habitEventTitleTV;
+
+    private Button delBtn;
+    private Button editConfirmBtn;
+    private TextInputEditText commentET;
     private TextView habitEventDateFormatTV;
-    private EditText habitEventCommentET;
+    private TextView habitEventTitleTV;
     private TextView habitEventLocationTV;
-    private ImageView habitEventImageView;
+    private EditText habitEventCommentET;
 
     private Habit habit;
     private HabitEvent hEvent;
 
-    public ViewHabitEventFragment() {
+    public EditHabitEventFragment() {
         // Required empty public constructor
     }
 
@@ -54,26 +56,25 @@ public class ViewHabitEventFragment extends Fragment {
         super.onCreate(savedInstanceState);
         setHasOptionsMenu(true);
 
-        habit = ViewHabitEventFragmentArgs.fromBundle(getArguments()).getHabit();
-        hEvent = ViewHabitEventFragmentArgs.fromBundle(getArguments()).getHabitEvent();
+        habit = EditHabitEventFragmentArgs.fromBundle(getArguments()).getHabit();
+        hEvent = EditHabitEventFragmentArgs.fromBundle(getArguments()).getHabitEvent();
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_view_habit_event, container, false);
 
-        // Initiate views
-        confirmHabitEventButton = view.findViewById(R.id.confirmHabitEventButton);
-        editHabitEventButton = view.findViewById(R.id.editHabitEventButton);
-        habitEventTitleTV = view.findViewById(R.id.viewHabitEventTitleTV);
-        habitEventDateFormatTV = view.findViewById(R.id.habitEventDateFormatTV);
-        habitEventCommentET = view.findViewById(R.id.habitEventCommentET);
-        habitEventLocationTV = view.findViewById(R.id.habitEventLocationTV);
-        habitEventImageView = view.findViewById(R.id.habitEventImage);
+        binding = FragmentEditHabitEventBinding.inflate(inflater, container, false);
 
-        return view;
+        delBtn = binding.deleteHabitEventButton;
+        editConfirmBtn = binding.confirmHabitEventButton;
+        commentET = binding.habitEventCommentET;
+        habitEventDateFormatTV = binding.habitEventDateFormatTV;
+        habitEventTitleTV = binding.editHabitEventTitleTV;
+        habitEventCommentET = binding.habitEventCommentET;
+        habitEventLocationTV = binding.habitEventLocationTV;
+
+        return binding.getRoot();
     }
 
     @Override
@@ -82,27 +83,17 @@ public class ViewHabitEventFragment extends Fragment {
 
         navController = Navigation.findNavController(view);
 
-        // confirm button clicked
-        confirmHabitEventButton.setOnClickListener(v -> {
-            NavDirections action = MobileNavigationDirections.actionGlobalNavigationTimeline(null);
-            navController.navigate(action);
-        });
-        // edit button
-        editHabitEventButton.setOnClickListener(v -> {
-            Log.d(TAG, "clicked edit habit event button");
-            NavDirections action = MobileNavigationDirections.actionGlobalEditHabitEventFragment(hEvent, habit);
-            navController.navigate(action);
-        });
+        editConfirmBtn.setOnClickListener(editConfirmOnClickListener);
+        delBtn.setOnClickListener(deleteOnClickListener);
 
-        setViewHabitEventFields();
+        setEditHabitEventFields();
     }
 
     /**
-     * Sets all the fields in view habit fragment
-     * to the habit event's info
+     * Sets the fields of the edit habit event fragment views
+     * to the proper values given by the habit event object
      */
-    private void setViewHabitEventFields() {
-
+    private void setEditHabitEventFields() {
         // Get date and set it to TextView
         String datePattern = "yyyy-MM-dd";
         SimpleDateFormat format = new SimpleDateFormat(datePattern, Locale.ROOT);
@@ -113,12 +104,13 @@ public class ViewHabitEventFragment extends Fragment {
         habitEventTitleTV.setText(hEvent.getTitle());
 
         habitEventCommentET.setText(hEvent.getComment());
+
         if (hEvent.getLocation() != null) {
             habitEventLocationTV.setText(hEvent.getLocation().toString());
         }
 
-        if (hEvent.getPhotoUrl() == null) {
-            habitEventImageView.setImageResource(R.color.transparent);
+        if (hEvent.getPhotoUrl() != null) {
+            // TODO set image view to the image given by habit event
         }
     }
 
@@ -145,4 +137,30 @@ public class ViewHabitEventFragment extends Fragment {
         super.onStop();
         MainActivity.showBottomNav(getActivity().findViewById(R.id.addHabitButton), getActivity().findViewById(R.id.extendBottomNav));
     }
+
+    private View.OnClickListener editConfirmOnClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View view) {
+
+            String comment = commentET.getText().toString();
+            // TODO get location, photograph from the user
+
+            hEvent.setComment(comment);
+
+            UserController.updateHabitEventInDb(hEvent, user -> {
+                NavDirections action = MobileNavigationDirections.actionGlobalNavigationTimeline(null);
+                navController.navigate(action);
+            });
+        }
+    };
+
+    private View.OnClickListener deleteOnClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View view) {
+            UserController.removeHabitEventFromDb(hEvent, user -> {
+                NavDirections action = MobileNavigationDirections.actionGlobalNavigationTimeline(null);
+                navController.navigate(action);
+            });
+        }
+    };
 }
