+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/timeline/TimelineListAdapter.java
@@ -11,14 +11,9 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import androidx.constraintlayout.widget.ConstraintLayout;
-
 import com.bumptech.glide.Glide;
 import com.cmput301f21t26.habittracker.R;
-import com.cmput301f21t26.habittracker.objects.Habit;
 import com.cmput301f21t26.habittracker.objects.HabitEvent;
-
-import org.w3c.dom.Text;
 
 import java.util.ArrayList;
