+++ code/app/src/main/res/layout/fragment_edit_habit.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.EditHabitFragment">
+    tools:context=".ui.habit.EditHabitFragment">
 
     <ScrollView
         android:layout_width="match_parent"
