+++ code/app/src/main/res/layout/fragment_edit_habit_event.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.EditHabitEventFragment">
+    tools:context=".ui.habitevent.EditHabitEventFragment">
 
     <ScrollView
         android:layout_width="match_parent"
