+++ code/app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/loginContainer"
-    tools:context=".fragments.LoginFragment">
+    tools:context=".ui.auth.LoginFragment">
 
     <TextView
         android:id="@+id/logTV"
