+++ code/app/src/main/res/layout/fragment_signup.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.SignupFragment">
+    tools:context=".ui.auth.SignupFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
