+++ code/app/src/main/res/layout/fragment_view_habit.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.ViewHabitFragment">
+    tools:context=".ui.habit.ViewHabitFragment">
 
     <ScrollView
         android:layout_width="match_parent"
