+++ code/app/src/main/res/layout/fragment_view_habit_event.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.ViewHabitEventFragment">
+    tools:context=".ui.habitevent.ViewHabitEventFragment">
 
     <ScrollView
         android:layout_width="match_parent"
