+++ code/app/src/main/res/menu/toolbar_menu.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
         <item
             android:id="@+id/action_settings"
             android:orderInCategory="100"
